Unit Class works correctly with new map, methods take a map object. BoardManager moved to Map/game_map GrassTile, WaterTile moved to Map/tile_types
