--- Map/GrassTile.py
@@ -1,9 +1,0 @@
-from Map.MapTile import MapTile
-
-
-class GrassTile(MapTile):
-	pass
-
-	tile_type = 'grass'
-	movement_mod = 1
-	color = (86, 176, 0)
