--- Map/WaterTile.py
@@ -1,9 +1,0 @@
-from Map.MapTile import MapTile
-
-
-class WaterTile(MapTile):
-	pass
-
-	tile_type = 'water'
-	movement_mod = 0
-	color = (0, 128, 255)
